Add jwt protected routes TODO: improve jwt implementation

+++ user-service/routes/user_routes.py
@@ -2,12 +2,11 @@ from fastapi import APIRouter, Depends, HTTPException
 from pydantic import UUID4
 from sqlalchemy.orm import Session
 from typing import List
-from schemas.user_schema import UserCreateSchema, UserUpdateSchema, UserResponseSchema
+from schemas.user_schema import UserCreateSchema, UserUpdateSchema, UserResponseSchema, UserLoginSchema
 from services.user_service import UserService
+from utils.dependencies import get_current_user
 from config.session import get_db
 
-
-# TODO: Password hashing and jwt impl
 
 router = APIRouter()
 user_service = UserService()
@@ -26,9 +25,14 @@ def get_user(user_id: UUID4, db: Session = Depends(get_db)):
     return user
 
 
-@router.post("/", response_model=UserCreateSchema)
-def create_user(user: UserCreateSchema, db: Session = Depends(get_db)):
+@router.post("/", response_model=UserResponseSchema)
+def register_user(user: UserCreateSchema, db: Session = Depends(get_db)):
     return user_service.create_user(db, user)
+
+
+@router.post("/login")
+def login(user: UserLoginSchema, db: Session = Depends(get_db)):
+    return user_service.login(db, user)
 
 
 @router.put("/{user_id}", response_model=UserResponseSchema)
@@ -37,7 +41,18 @@ def update_user(user_id: UUID4, user: UserUpdateSchema, db: Session = Depends(ge
 
 
 @router.delete("/{user_id}")
-def delete_user(user_id: UUID4, db: Session = Depends(get_db)):
+def delete_user(
+    user_id: UUID4,
+    db: Session = Depends(get_db),
+    current_user: UserResponseSchema = Depends(get_current_user)
+):
+    if str(current_user.user_id) != str(user_id) and current_user.role != "admin":
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to delete this user."
+        )
+
+    user_service = UserService()
     user_service.delete_user(db, user_id)
     return {"message": "User deleted successfully"}
